# openapi.yaml
openapi: 3.0.3
info:
  title: WinOJ API
  description: |
    Windows 在线评测系统 API。支持多语言代码评测、安全沙箱执行、自定义计分脚本、比赛管理、文章系统、Rating 机制、AI 代码安全审查等功能。

    ## 认证
    大部分接口需携带 JWT Token：`Authorization: Bearer <access_token>`
    Access Token 有效期 15 分钟，过期后通过 `POST /auth/refresh` 刷新。

    ## 错误响应
    ```json
    { "code": 1|2|3|4|5|6, "reason": "ERR_xxx", "message": "..." }
    ```
    错误码：1=参数错误, 2=状态错误, 3=未找到, 4=限流, 5=未认证, 6=权限不足

    ## 角色层级
    user < teacher < admin < su
  version: 1.6.0
  contact:
    name: WinOJ
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /contests:
    get:
      tags: [比赛]
      summary: 比赛列表
      responses:
        '200':
          description: 比赛列表
    post:
      tags: [比赛]
      summary: 创建比赛
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, start_time, end_time]
              properties:
                title:
                  type: string
                description:
                  type: string
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                is_virtual:
                  type: integer
      responses:
        '201':
          description: 创建成功
  /contests/{id}:
    get:
      tags: [比赛]
      summary: 比赛详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 比赛详情
    put:
      tags: [比赛]
      summary: 更新比赛
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [比赛]
      summary: 删除比赛
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
  /contests/{id}/problems:
    post:
      tags: [比赛]
      summary: 添加题目到比赛
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [problem_id]
              properties:
                problem_id:
                  type: integer
                alias:
                  type: string
                sort_order:
                  type: integer
                  default: 0
      responses:
        '201':
          description: 添加成功
  /contests/{id}/problems/{pid}:
    delete:
      tags: [比赛]
      summary: 从比赛中移除题目
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: pid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 移除成功
  /contests/{id}/invite:
    post:
      tags: [比赛]
      summary: 邀请用户参赛
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: integer
      responses:
        '200':
          description: 邀请成功
  /contests/{id}/join:
    post:
      tags: [比赛]
      summary: 加入比赛
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 加入成功
  /contests/{id}/participants:
    get:
      tags: [比赛]
      summary: 参赛用户列表
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 参赛用户列表
  /contests/{id}/participants/{uid}:
    delete:
      tags: [比赛]
      summary: 移除参赛用户
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: uid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 移除成功
  /contests/{id}/leaderboard:
    get:
      tags: [比赛]
      summary: 比赛排行榜
      description: 可选认证。参赛者/teacher/admin/su 返回完整排行榜（含 rank、total_score、total_time、各题得分）；非参赛者或未登录仅返回冻结状态（leaderboard 为空），不泄露排名。冻结时只统计冻结时刻前的提交。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 排行榜
  /contests/{id}/unfreeze:
    post:
      tags: [比赛]
      summary: 解除排行榜冻结
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 解冻成功
  /contests/{id}/announcements:
    get:
      tags: [比赛]
      summary: 比赛内公告列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 公告列表
  /contests/{id}/virtual-start:
    post:
      tags: [比赛]
      summary: 发起虚拟参赛
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 虚拟参赛已启动
  /contests/{id}/plagiarism-check:
    post:
      tags: [比赛]
      summary: 对比赛所有题目一键查重
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                threshold:
                  type: integer
                  default: 80
                min_submissions:
                  type: integer
                  default: 2
      responses:
        '200':
          description: 查重任务已启动
  /contests/{id}/discussions:
    get:
      tags: [比赛]
      summary: 比赛讨论列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 讨论列表

  # ===== ARTICLES =====
...